requirements.txt
Django==5.0
tensorflow==2.17.1
keras==3.5.0
numpy==1.26.4
pandas==2.2.2
matplotlib
opencv-python
ffmpeg-python
injector
